Extract `MarkerHoverParticipant` to its own file
